Moved sample image generation to its own file.
